worker/requirements.txt
# Minimal worker-specific dependencies (core app deps live at repo root requirements.txt)
azure-identity>=1.19.0
azure-storage-blob>=12.20.0
azure-servicebus>=7.12.0
pydantic>=2.9.0
